Fix for: An empty record saved to smartstore doesn't get a row in the FTS table
Impact:
When records is updated, it still doesn't have a row in the FTS table
As a result, record cannot be found with a FTS query.
Also when if more non-empty records are created, their id won't match the record ids in the main soup table and so FTS query returns the wrong records.

How did we find the bug:
SmartSyncExplorerHybrid add creates an empty record before going into the Edit screen.
If you back out from the screen, the record is deleted, if you save it gets updated

Because all FTS fields were empty, no row were created in the FTS table.
When the save happens, there is no row to update.
As a result, those records never show up in the search screen (until the records are synched up and the app is restarted and a sync down bring the data back).

Fix:
Pass the id in the values so that a row actually gets created.

How come we didn't see the bug before?
Existing code works for non-empty record, the id column is auto-filled, and ends up with the same id as the corresponding row in the main soup table.
It's only when no values are passed through that the row is not created.