Major refactor to view state handling
The coupled VMs for the list and detail panes was growing cumbersome as expected, so this refactors
the UI state change handling by architecting a faux FSM which delegates UI state transition
decisions to the states themselves.  Then the Activity VM can make higher-level decisions when it
knows an event should affect more than one pane.