Using prompt=login during advanced authentication so that user must re-enter credentials
Previously if a user did a logout followed by a login, they would go straight to the allow/deny screen.
Changes were made initially in the internal fork of the Mobile SDK (see [here](https://git.soma.salesforce.com/Android/SalesforceMobileSDK-Android-Public/pull/283)).
There is a new property shareBrowserSession (and a method setBrowserLoginEnabled) on SalesforceSDKManager to go back to the old behavior.
That flag can also be set through the AuthConfig.