Not getting user info to refresh access token
User info is a /services/oauth2 end point, not /services/data one.
It returns a 403 for an invalid access token which we don't handle automatically. See end of that [doc](https://help.salesforce.com/s/articleView?language=en_US&id=sf.remoteaccess_using_userinfo_endpoint.htm&type=5).
Initially, we would refresh the access token upon getting a 401 or 403, but certain clients relied on that 403 to determinate if their token was invalid, so we had it controlled by a flag, and then we got rid of the flag but we kept using the calls to user info for our "cheap rest call" to refresh the access token.

In this PR, we are adding a wrapper for "limits" (with a test).
And we are now using a call to "limits" as our "cheap rest call" (in the IDP code and in SalesforceDroidGapActivity)